part of comments fixed
