--- components/Loader/Loader.css
@@ -1,43 +1,0 @@
-.loader_size-s {
-    display: inline-block;
-    width: 20px;
-    height: 20px;
-    border: 3px solid red;
-    border-right-color: #f9f9f9;
-    border-radius: 100%;
-    animation: spin 1s linear infinite;
-    transform: rotate(35deg);
-}
-
-
-
-.loader_size-m {
-    display: inline-block;
-    width: 40px;
-    height: 40px;
-    border: 5px solid red;
-    border-right-color: #f9f9f9;
-    border-radius: 100%;
-    animation: spin 1s linear infinite;
-    transform: rotate(35deg);
-}
-
-.loader_size-l {
-    display: inline-block;
-    width: 60px;
-    height: 60px;
-    border: 6px solid red;
-    border-right-color: #f9f9f9;
-    border-radius: 100%;
-    animation: spin 1s linear infinite;
-    transform: rotate(35deg);
-}
-
-@keyframes spin {
-    from {
-        transform: rotate(0deg);
-    }
-    to {
-        transform: rotate(360deg);
-    }
-}
