--- components/WithLoader/WithLoader.tsx
@@ -1,21 +1,0 @@
-import React from "react";
-
-import { Loader } from "../Loader/Loader";
-import { LoaderProps } from "../Loader/Loader";
-
-export type WithLoaderProps = React.PropsWithChildren<{
-  loading: boolean;
-  children?: React.ReactNode;
-}>;
-
-export const WithLoader: React.FC<WithLoaderProps> = ({
-  loading,
-  children,
-}) => {
-  return (
-    <div>
-      {children}
-      {loading && <Loader loading={loading} />}
-    </div>
-  );
-};
