--- src/config/api.ts
@@ -1,1 +1,0 @@
-export const API_ENDPOINTS = {}
