+++ src/config/routes.ts
@@ -1,0 +1,4 @@
+export const APP_ROUTES = {
+  RECIPE: '/',
+  RECIPE_DETAIL: '/detail-recipe/:id',
+}
