+++ src/pages/DetailRecipePage/DetailRecipePage.tsx
@@ -1,12 +1,12 @@
 import React, { useEffect } from 'react'
 
+import { APP_ROUTES } from '@config/routes'
 import heart from '@img/Heart.svg'
 import path from '@img/Path.svg'
 import star from '@img/Star 3.svg'
+import { useRecipeContext } from '@pages/RecipeContext/RecipeContex'
 import { Link, useParams } from 'react-router-dom'
 
-// import RES from '../../config/result'
-import { useRecipeContext } from '../RecipesCardsPage/components/RecipeContex/RecipeContex'
 import styles from './DetailRecipePage.module.scss'
 
 const DetailRecipePage = () => {
@@ -19,17 +19,12 @@ const DetailRecipePage = () => {
 
   return (
     <div className={styles.detailed_block}>
-      <Link to={`/`}>
-        <button className={styles.next_button}>
+      <Link to={APP_ROUTES.RECIPE}>
+        <button className={styles.back_button}>
           <img src={path} alt={'path'} />
         </button>
       </Link>
-      {/*<img*/}
-      {/*  src={'https://spoonacular.com/recipeImages/716429-556x370.jpg'}*/}
-      {/*  alt={'bla'}*/}
-      {/*  width='100%'*/}
-      {/*/>*/}
-      <img src={detailsInfo?.image} alt={detailsInfo?.title} width="100%" />
+      <img src={detailsInfo?.image} alt={detailsInfo?.title} width='100%' />
       <div className={styles.description}>
         <div className={styles.title}>{detailsInfo?.title}</div>
         <div className={styles.rating_block}>
@@ -37,14 +32,12 @@ const DetailRecipePage = () => {
             <img
               src={star}
               alt={'star'}
-              width="16px"
-              height="16px"
-              className={styles.img_star}
+              className={styles['stars_block__img']}
             />
             <p>3.8 Rating</p>
           </div>
           <div className={styles.heart_block}>
-            <img src={heart} alt={'heart'} className={styles.img_heart} />
+            <img src={heart} alt={'heart'} className={styles['heart_block__img']} />
             <p>3.8 Rating</p>
           </div>
         </div>
