+++ src/pages/MenuPage/components/MultiDropdown/MultiDropdown.module.scss
@@ -1,0 +1,28 @@
+@import 'src/styles/variables';
+
+.dropdown {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  margin-top: 16px;
+  margin-bottom: 20px;
+}
+
+.dropdown_button {
+  border: transparent;
+  background: linear-gradient(98.81deg, $light-pink -0.82%, $dark-pink 101.53%);
+  border-radius: 7px;
+  text-align: left;
+  @include button-size(176px, 36px);
+}
+
+.dropdown_item {
+  display: block;
+  width: 176px;
+  height: 36px;
+  border: transparent;
+  background: linear-gradient(98.81deg, $light-pink -0.82%, $dark-pink 101.53%);
+  mix-blend-mode: soft-light;
+  border-radius: 7px;
+  text-align: left;
+}
