+++ src/pages/MenuPage/components/RecipeCardItem/RecipeCardItem.tsx
@@ -3,6 +3,7 @@ import React from 'react'
 import star from '@img/Star 3.svg'
 import { Link } from 'react-router-dom'
 
+
 import styles from './RecipeCardItem.module.scss'
 
 type CardProps = {
@@ -15,16 +16,13 @@ type CardProps = {
 }
 
 export const RecipeCardItem: React.FC<CardProps> = ({
-  meal,
-  image,
-  title,
-  subtitle,
-  calories,
-  onClick,
-}) => {
-  // const [hasRender, setRender] = useState(false)
-  // const onShow = React.useCallback(() => setRender(true), [])
-
+                                                      meal,
+                                                      image,
+                                                      title,
+                                                      subtitle,
+                                                      calories,
+                                                      onClick,
+                                                    }) => {
   return (
     <div className={styles.card} onClick={onClick}>
       <div className={styles.card_block}>
@@ -34,8 +32,8 @@ export const RecipeCardItem: React.FC<CardProps> = ({
         </div>
         <img
           className={styles.card_img}
-          width="87px"
-          height="70px"
+          width='87px'
+          height='70px'
           src={image}
           alt={'food'}
         />
