--- src/pages/RecipesCardsPage/components/Input/Input.module.scss
@@ -1,23 +1,0 @@
-.input_wrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-end;
-    position: relative;
-}
-.input {
-    width: 70%;
-    height: 36px;
-    padding-left: 13%;
-    background: linear-gradient(98.81deg, #FFF0F0 -0.82%, #FFDFDF 101.53%);
-    border-radius: 7px;
-    font-weight: 400;
-    font-size: 12px;
-    line-height: 28px;
-    border-color: transparent;
-}
-
-.input_img {
-    position: absolute;
-    top: 25%;
-    left: 20%;
-}
