--- src/pages/RecipesCardsPage/components/Menu/index.tsx
@@ -1,5 +1,0 @@
-export { default } from './Menu'
-
-//внутри components только компоненты этой страницы
-
-//Директория src/components — где размещать компоненты, которые используются по ВСЕМУ проекту
